requirements.txt
requests
pandas
python-dotenv
google-cloud-bigquery
pyarrow
